Finished all models except Prescriptions
Redo the design of the site a little and add the circle div on every page.

+++ Medik_01/Medik/Models/Examination.cs
@@ -9,10 +9,10 @@ namespace Medik.Models
         private DateTime dateOfExam;
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Required]
-        public int PatientId { get; set; }
+        public long PatientId { get; set; }
 
         [Required]
         [Display(Name = "Date of examination")]
